refactor: script de verificar se existe pessoas na tabela com o id passado

+++ Back-End/cipa/src/main/java/br/edu/ifpe/cipa/service/PessoaService.java
@@ -1,6 +1,5 @@
 package br.edu.ifpe.cipa.service;
 
-import java.util.Iterator;
 import java.util.List;
 
 import org.springframework.stereotype.Service;
@@ -14,32 +13,45 @@ public class PessoaService {
 	
 	PessoaDao pessoadao = new PessoaDao();
 	
+	public boolean entidadeJaExiste(Pessoa pessoa) {
+		return pessoadao.listar().stream().anyMatch(p -> p.getEmail().equals(pessoa.getEmail()));
+	}
+
+	public boolean entidadeJaExisteId(Integer id) {
+		return pessoadao.listar().stream().anyMatch(p -> p.getId() == id);
+	}
+
 	public List<Pessoa> listar() {
 		return pessoadao.listar();
 	}
+
+	public Pessoa listarOne(Integer id) throws Exception {
+		if (!this.entidadeJaExisteId(id)) {
+			throw new Exception("Id não encontrado");
+		}
+		return pessoadao.listar().get(id - 1);
+	}
 	
-	public void alterar(Pessoa pessoa) {
+	public void alterar(Pessoa pessoa) throws Exception {
+		if (!this.entidadeJaExiste(pessoa)) {
+			throw new Exception("Pessoa não existe");
+		}
 		pessoadao.alterar(pessoa);
 	}
+
 	
-	public boolean inserir(Pessoa pessoa) {
-		List<Pessoa> allPessoas = pessoadao.listar();
-		var value = false;
-		for (Iterator<Pessoa> iterator = allPessoas.iterator(); iterator.hasNext(); ) { 
-			Pessoa p = iterator.next(); 
-			if (p.getEmail().equals(pessoa.getEmail())) {
-				value = true;
-			}
+	public void inserir(Pessoa pessoa) throws Exception {
+		if (this.entidadeJaExiste(pessoa)) {
+			throw new Exception("Pessoa ja cadastrado");
 		}
-		if(value == false) {
-			pessoadao.inserir(pessoa);
-			return true;
-		} else {
-			return false;
-		}
+		pessoadao.inserir(pessoa);
+	
 	}
 	
-	public void remover(int id) {
+	public void remover(int id) throws Exception {
+		if (!this.entidadeJaExisteId(id)) {
+			throw new Exception("Pessoa não existe");
+		}
 		pessoadao.remover(id);
 	}
